Havent pushed in a while combinated of few days of work + working on the dissertation
07/04/2022
- Had Final Meeting with lecturer
- Worked on 3/4 of the dissertation
- Fixed Bugs on the enemies not being removed from the list
- Worked on bug that team mates can do 2 moves if mana goes under whilst doing a move
08/04/2022
- Mess around with the balance scaling of the game
- Made it so level and level 5 enemies utilize either 2 or less enemies
- Created Level 4
- Created Level 5 aka final boss
- Realised that teammate can still do double attacks, created a bool to make sure that this doesnt happen
- The player had a few bugs where they were also breaking the game with mana going under the requirement.
- Created materials to give a different feeling for each level aside from those already done and level 1
- Got a couple of friends to test out the game at its current state.

TEST BUILD NOTES

BUTTONS FOR TARGETS ALL OVER THE PLACE

HEAL NAMES ARENT DONE FOR SOME

ENEMIES EARNING SANITY VIA ATTACKING???

ENEMIES ALSO HAVE BUG FOR SPELLS

PLAYER BREAKS GAME IF HE DOESNT HAVE SANITY AND U PRESS SPELL THEN PRESS ATTACK

09/04/2022 (Early Early Morning)
- Working on fixing few of the bugs found from the early early testing of the game

+++ Emotional_Sanity_FMP/Assets/Scripts/BaseEntities.cs
@@ -72,12 +72,12 @@ public class BaseEntities : MonoBehaviour
                     //Debug.Log(HP);
                 } else if (SP >= 75)
                 {
-                    HP -= ((damage * (int)strength) / (int)weapondefence) * 0.7f;
+                    HP -= (int)((damage * (int)strength) / (int)weapondefence) * 0.7f;
                     Debug.Log("Weapon Effective! + YOUR SANITY IS HIGH THEREFORE YOU TOOK LESS DAMAGE THAN USUAL");
                     //Debug.Log(HP);
                 } else
                 {
-                    HP -= (damage * (int)strength) / (int)weapondefence;
+                    HP -= (int)(damage * (int)strength) / (int)weapondefence;
                     Debug.Log("Weapon Effective!");
                 }
             }
@@ -86,7 +86,7 @@ public class BaseEntities : MonoBehaviour
             {
                 if (SP <= 25)
                 {
-                    HP -= (((damage * (int)strength) / (int)weapondefence) * 75 / 100) * 1.5f;
+                    HP -= (int)(((damage * (int)strength) / (int)weapondefence) * 75 / 100) * 1.5f;
                     if (SP < 0)
                     {
                         SP = 0;
@@ -96,13 +96,13 @@ public class BaseEntities : MonoBehaviour
                 }
                 else if (SP >= 75)
                 {
-                    HP -= (((damage * (int)strength) / (int)weapondefence) * 75 / 100) * 0.7f;
+                    HP -= (int)(((damage * (int)strength) / (int)weapondefence) * 75 / 100) * 0.7f;
                     Debug.Log("Blocked!: Weapon Effective! + YOUR SANITY IS HIGH THEREFORE YOU TOOK LESS DAMAGE THAN USUAL");
                     //Debug.Log(HP);
                 }
                 else
                 {
-                    HP -= ((damage * (int)strength) / (int)weapondefence) * 75 / 100;
+                    HP -= (int)((damage * (int)strength) / (int)weapondefence) * 75 / 100;
                     Debug.Log("Blocked!: Weapon Effective!");
                 }
